wip: refactored code etc.
