+++ src/app/(auth)/template.js
@@ -28,7 +28,7 @@ const AuthTemplate = ({ children }) => {
         </div>
         <Dots className="absolute -left-48 -bottom-60 md:-left-40 md:-bottom-40 text-queen-orange" />
       </div>
-      <div className="bg-queen-white flex justify-center items-center h-full">
+      <div className="flex justify-center items-center h-full">
         <div className="w-full max-w-sm mx-auto space-y-8 px-8 md:px-6">
           {children}
         </div>
