+++ src/app/(dashboard)/(creator)/opportunities/[slug]/page.js
@@ -68,7 +68,6 @@ export default async function Opportunity({ params }) {
 
   return (
     <div className="bg-purple-dots bg-repeat-x bg-[center_bottom]">
-      <MainNav />
       <Container size="2xl">
         <div className="mt-20 pb-60 space-y-8">
           <div>
