+++ src/app/(dashboard)/template.js
@@ -1,0 +1,10 @@
+import MainNav from "@/components/MainNav";
+
+const Template = ({ children }) => (
+  <>
+    <MainNav />
+    {children}
+  </>
+);
+
+export default Template;
