+++ src/app/layout.js
@@ -25,7 +25,7 @@ export default function RootLayout({ children }) {
   return (
     <html
       lang="en"
-      className={`${formaDJRBanner.variable} ${formaDJRDeck.variable}`}
+      className={`bg-queen-white ${formaDJRBanner.variable} ${formaDJRDeck.variable}`}
     >
       <body>
         {children}
