+++ src/components/Empty.jsx
@@ -1,0 +1,3 @@
+const Empty = () => <div></div>;
+
+export default Empty;
